Move update API key types to server types package
