+++ x-pack/plugins/security/server/routes/api_keys/update.ts
@@ -5,95 +5,25 @@
  * 2.0.
  */
 
-import type { estypes } from '@elastic/elasticsearch';
-
 import { schema } from '@kbn/config-schema';
-import type { TypeOf } from '@kbn/config-schema';
-import { elasticsearchRoleSchema, getKibanaRoleSchema } from '@kbn/security-plugin-types-server';
+import type { UpdateAPIKeyResult } from '@kbn/security-plugin-types-server';
+import {
+  getUpdateRestApiKeyWithKibanaPrivilegesSchema,
+  updateCrossClusterApiKeySchema,
+  updateRestApiKeySchema,
+} from '@kbn/security-plugin-types-server';
 
 import type { RouteDefinitionParams } from '..';
 import { UpdateApiKeyValidationError } from '../../authentication/api_keys/api_keys';
 import { wrapIntoCustomErrorResponse } from '../../errors';
 import { createLicensedRouteHandler } from '../licensed_route_handler';
 
-/**
- * Response of Kibana Update API key endpoint.
- */
-export type UpdateAPIKeyResult = estypes.SecurityUpdateApiKeyResponse;
-
-/**
- * Request body of Kibana Update API key endpoint.
- */
-export type UpdateAPIKeyParams =
-  | UpdateRestAPIKeyParams
-  | UpdateCrossClusterAPIKeyParams
-  | UpdateRestAPIKeyWithKibanaPrivilegesParams;
-
-export type UpdateRestAPIKeyParams = TypeOf<typeof restApiKeySchema>;
-export type UpdateCrossClusterAPIKeyParams = TypeOf<typeof crossClusterApiKeySchema>;
-export type UpdateRestAPIKeyWithKibanaPrivilegesParams = TypeOf<
-  ReturnType<typeof getRestApiKeyWithKibanaPrivilegesSchema>
->;
-
-const restApiKeySchema = schema.object({
-  type: schema.maybe(schema.literal('rest')),
-  id: schema.string(),
-  role_descriptors: schema.recordOf(schema.string(), schema.object({}, { unknowns: 'allow' }), {
-    defaultValue: {},
-  }),
-  metadata: schema.maybe(schema.object({}, { unknowns: 'allow' })),
-});
-
-const crossClusterApiKeySchema = restApiKeySchema.extends({
-  type: schema.literal('cross_cluster'),
-  role_descriptors: null,
-  access: schema.object(
-    {
-      search: schema.maybe(
-        schema.arrayOf(
-          schema.object(
-            {
-              names: schema.arrayOf(schema.string()),
-            },
-            { unknowns: 'allow' }
-          )
-        )
-      ),
-      replication: schema.maybe(
-        schema.arrayOf(
-          schema.object(
-            {
-              names: schema.arrayOf(schema.string()),
-            },
-            { unknowns: 'allow' }
-          )
-        )
-      ),
-    },
-    { unknowns: 'allow' }
-  ),
-});
-
-const getRestApiKeyWithKibanaPrivilegesSchema = (
-  getBasePrivilegeNames: Parameters<typeof getKibanaRoleSchema>[0]
-) =>
-  restApiKeySchema.extends({
-    role_descriptors: null,
-    kibana_role_descriptors: schema.recordOf(
-      schema.string(),
-      schema.object({
-        elasticsearch: elasticsearchRoleSchema.extends({}, { unknowns: 'allow' }),
-        kibana: getKibanaRoleSchema(getBasePrivilegeNames),
-      })
-    ),
-  });
-
 export function defineUpdateApiKeyRoutes({
   router,
   authz,
   getAuthenticationService,
 }: RouteDefinitionParams) {
-  const bodySchemaWithKibanaPrivileges = getRestApiKeyWithKibanaPrivilegesSchema(() => {
+  const bodySchemaWithKibanaPrivileges = getUpdateRestApiKeyWithKibanaPrivilegesSchema(() => {
     const privileges = authz.privileges.get();
     return {
       global: Object.keys(privileges.global),
@@ -106,8 +36,8 @@ export function defineUpdateApiKeyRoutes({
       path: '/internal/security/api_key',
       validate: {
         body: schema.oneOf([
-          restApiKeySchema,
-          crossClusterApiKeySchema,
+          updateRestApiKeySchema,
+          updateCrossClusterApiKeySchema,
           bodySchemaWithKibanaPrivileges,
         ]),
       },
